// src/modules/users/models/AddressView.ts
export interface AddressView {
    name?: string;
    email?: string;
    mobile: string;
    flat: string;
    landmark: string;
    street: string;
    state: string;
    city: string;
    country: string;
    pinCode: string;
    userObj?: string;
    _id?: string;
    createdAt?: string;
    updatedAt?: string;
    __v?: number;
}